backend/pyproject.toml
[build-system]
requires = [
    "setuptools>=68",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "appwrite>=13.0.0",
    "beam-client>=0.2.186",
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=3.13",
    "flask-socketio>=5.5.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "muspy>=0.5.0",
    "numpy>=2.3.3",
    "protobuf>=4.25.8,<5",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "scipy>=1.16.2",
    "setuptools>=80.9.0",
    "statsmodels>=0.14.5",
    "werkzeug>=3.1.3",
]

[project.scripts]
note-backend = "app.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
